Create Links-to-chat-generating-project-files.txt
This file is going to be modified heavily soon with full documentation of the development process.  You'll be able to show these to your AI companions to catch them up on exactly what this is all about and then they can in turn catch you up on what exactly it's all about.